feat: add task prerequisites support
